Chapter11 of Laravel 5.4 Email Verification
In this chapter the verify system is established. We change the registration login (need to change the database) with a column named 'status'. New users need to activate the account through the link to sent to their email boxes after registration, otherwise they will fail in login. Go and review the readme file inside the project as well as the code.